move address components into the receive section
