+++ src/components/TransactionModal/styles.ts
@@ -1,5 +1,5 @@
 import * as m from 'motion/react-m';
-import styled, { css } from 'styled-components';
+import styled from 'styled-components';
 
 export const Wrapper = styled('div')`
     position: fixed;
