+++ src/components/transactions/WalletSidebarContent.styles.ts
@@ -1,4 +1,3 @@
-import { m } from 'motion/react';
 import styled from 'styled-components';
 
 export const WalletSections = styled.div`
