+++ src/components/transactions/receive/CopyAddress.tsx
@@ -1,7 +1,7 @@
 import { useWalletStore } from '@app/store';
 import { useCopyToClipboard } from '@app/hooks';
 import { useTranslation } from 'react-i18next';
-import { CopyAddressButton } from '../../wallet/Address.style';
+import { CopyAddressButton } from './Address.style';
 
 export function CopyAddress({ useEmoji }: { useEmoji: boolean }) {
     const { t } = useTranslation('wallet');
