+++ src/components/transactions/receive/Receive.tsx
@@ -1,6 +1,6 @@
 import { Wrapper } from './receive.styles';
-import { AddressQRCode } from '@app/components/wallet/AddressQRCode';
-import { CopyAddress } from '@app/components/wallet/CopyAddress';
+import { AddressQRCode } from './AddressQRCode';
+import { CopyAddress } from './CopyAddress';
 import { useState } from 'react';
 
 export function Receive() {
